// src/visual-dashboard-frontend/src/interfaces/schemasReducer.ts
export interface TabInList {
    id: number
    name: string,
    sensors_ids: any[]
}
// ...
export interface PublishedSchema {
    id: number,
    name: string,
    schema_tabs: TabInList[]
}

export interface Location {
    id: number,
    title: string,
    published_schema?: PublishedSchema
    devices: any
}